mycfg: support "vim" mode
Basically changes the colorscheme to "vim", and avoids loading some
extensions that don't play nice with the "vim" colorscheme.

"vim" mode shall be activated if neovim is invoked as vim.
ie. assume this symlink exists -

	~/.local/bin/vim  ->  /usr/bin/nvim

Now, if ~/.local/bin is in $PATH, then running "vim" at the prompt shall
execute /usr/bin/nvim, but with argv[0] set to "vim" instead.

This behaviour was broken initially. This has been fixed in this commit -
https://github.com/neovim/neovim/commit/0ce065a332cbd3874bd68a86fa9eda20c5467170